alpha port
git-svn-id: https://gforge.info.ucl.ac.be/svn/mozart@811 ada56829-ad1f-0410-b00f-83cda6628aec

+++ trunk/mozart/platform/emulator/taskstk.hh
@@ -77,7 +77,7 @@ typedef StackEntry TaskStackEntry;
 
 
 /* if you ever change this check POPTASK */
-const TaskStackEntry emptyTaskStackEntry = (TaskStackEntry)-1;
+const TaskStackEntry emptyTaskStackEntry = (TaskStackEntry)(unsigned int32)-1;
 
 class TaskStack: public Stack {
 public:
